dp: introduced a basic pagination extension
